Grading
I don't mind using in class examples, but in the future let's at least use a different dataset! That's on me in the prompt.

I've said this to everyone, but in the future have your Git commits mention the work you did since the last commit.

Check my comment on the UI. I duct you a few points there.